% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_user.R
\name{authenticate_user}
\alias{authenticate_user}
\title{Authenticate User}
\usage{
authenticate_user(client_id = Sys.getenv("X_CLIENT_ID"))
}
\arguments{
\item{client_id}{A string containing the OAuth 2.0 client ID for authenticating
with the X API. By default, this argument retrieves the token from the
environment variable \code{X_CLIENT_ID} (via \code{Sys.getenv("X_CLIENT_ID")}).
Adding your client ID to your \code{.Renviron} file ensures it is securely
stored and accessible without needing to manually input it for each
session.}
}
\value{
A list containing the token
}
\description{
Authenticates with the X API using OAuth 2.0 Authorization Code Flow with offline access.
Falls back to re-authentication if cached token fails.
}
\keyword{internal}
