% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR.R
\name{bioc.upgrade}
\alias{bioc.upgrade}
\title{bioc.update}
\usage{
bioc.upgrade()
}
\value{
no return value
}
\description{
Update Bioconductor packages in your system
}
\details{
Running this function will automatically upgrade the installed bioconductor packages in your system.
}
\author{
ZG Zhao
}
