% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR.R
\name{calHTqPCR}
\alias{calHTqPCR}
\title{Compare gene expression of sample pairs by limma method: FC and stats.}
\usage{
calHTqPCR(
  dt,
  ref.gene,
  comp.strings = NULL,
  comp.type = 1,
  norm.method = "deltaCt",
  verbose = FALSE
)
}
\arguments{
\item{dt}{Data of qPCRset class, read with readDataHTqPCR function. Please refer to readDataHTqPCR function info.}

\item{ref.gene}{String. Names of reference gene(s).}

\item{comp.strings}{String of length 1. Compare stirng related to sample name(s).}

\item{comp.type}{Integer. 1: single control; 2: paired compares.}

\item{norm.method}{String. Default: del}

\item{verbose}{pass to \code{\link{HTqPCR::normalizeCtData}}}
}
\value{
result: results of limmaCtData (list); contrast: contranst sample labels (named strings); d.norm: result of normalizeCtData.
}
\description{
Calculate and return major results of HTqPCR.
}
\details{
Return results includes d.norm, fold change and stats table.
}
\author{
ZG Zhao
}
