% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-tags.r
\name{calTags}
\alias{calTags}
\title{calTags function}
\usage{
calTags(
  path = ".",
  pattern = "\\\\.[RrSs]$",
  recursive = FALSE,
  src = list.files(path = path, pattern = pattern, full.names = TRUE, recursive =
    recursive),
  cal.type = "fun",
  keep.re = NULL,
  verbose = getOption("verbose")
)
}
\arguments{
\item{path}{}

\item{pattern}{}

\item{recursive}{}

\item{src}{}

\item{cal.type}{String 'fun' or 'class'}

\item{keep.re}{}

\item{verbose}{}
}
\description{
Calculate function or class number in source files
}
\author{
ZG Zhao
}
