% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_from_x.R
\name{delete_from_x}
\alias{delete_from_x}
\title{Delete a Post on X}
\usage{
delete_from_x(post_id)
}
\arguments{
\item{post_id}{The ID of the post you want to delete.}
}
\description{
Delete specified Post (in the path) by ID via the \href{https://docs.x.com/x-api/posts/post-delete-by-post-id}{delete a post endpoint}.
}
\examples{
\dontrun{
delete_from_x(post_id = targeted_post_id)
}
}
