% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_env.R
\docType{data}
\name{.x_env}
\alias{.x_env}
\title{Internal Environment for Caching Package State}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.x_env
}
\description{
This environment is used to store temporary state such as OAuth tokens,
runtime flags, or other reusable objects during the session.
It is not exported and should not be accessed directly by users.
}
\keyword{internal}
