% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbar.R
\name{errorbar}
\alias{errorbar}
\title{errorbar() function}
\usage{
errorbar(x, y, sd.upr, sd.lwr, horiz = FALSE, cex = 1, lwd = 1, col = 1)
}
\arguments{
\item{x}{}

\item{y}{}

\item{sd.upr}{}

\item{sd.lwr}{}

\item{horiz}{}

\item{cex}{scale factor for blunt end}

\item{lwd}{}

\item{col}{}
}
\value{
This fuction works as side effects.
}
\description{
Plot error bars for scatter, line or barplots with arrows() function.
}
\author{
ZG Zhao
}
