% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etags.r
\name{etags}
\alias{etags}
\title{etags function}
\usage{
etags(
  path = ".",
  pattern = "\\\\.el$",
  recursive = TRUE,
  src = list.files(path = path, pattern = pattern, full.names = TRUE, recursive =
    recursive, ignore.case = TRUE),
  odir = getwd(),
  append = FALSE
)
}
\arguments{
\item{path}{Path to source files}

\item{pattern}{Source file subfix pattern}

\item{recursive}{(TRUE, logical)}

\item{src}{Vector of source file names}

\item{odir}{Path to output TAGS file}

\item{append}{(FALSE, logical)}
}
\description{
A R warper for etags program
}
\examples{
require(xtools)
etags("/usr/local/share/emacs", pattern = "\\\\.el",
      odir="~/.emacs.d", append=FALSE)
}
\author{
ZGUANG
}
