% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_post.R
\name{extract_post}
\alias{extract_post}
\title{Extract Post Data from Timeline}
\usage{
extract_post(timeline, tz = Sys.timezone(), include_referenced_posts = TRUE)
}
\arguments{
\item{timeline}{A list containing the timeline data retrieved from the X API.}

\item{include_referenced_posts}{Logical. Whether to include referenced posts in the output. Defaults to TRUE.}
}
\value{
A tibble containing structured post data.
}
\description{
Processes the timeline data retrieved from the X API to wrangle post data,
including metadata such as likes, retweets, replies, and impressions.
}
\examples{
\dontrun{
timeline <- get_timeline(
  username = "XDevelopers",
  max_results = 100,
  start_time = iso_8601(Sys.Date() - 7)
)
post <- extract_post(timeline)
}
}
