% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_user.R
\name{extract_user}
\alias{extract_user}
\title{Extract User Data from Timeline}
\usage{
extract_user(timeline)
}
\arguments{
\item{timeline}{A list containing the timeline data retrieved from the X API.}
}
\value{
A tibble containing structured user data.
}
\description{
Processes the timeline data retrieved from the X API to extract user metadata,
including profile details and public metrics.
}
\examples{
\dontrun{
timeline <- get_timeline(
  username    = "XDevelopers",
  max_results = 100,
  start_time  = iso_8601(Sys.Date() - 7)
)
user <- extract_user(timeline)
}
}
