% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gap-barplot.R
\name{gap.barplot}
\alias{gap.barplot}
\title{gap.barplot function}
\usage{
gap.barplot(df, y.cols = 1:ncol(df), sdu.cols = NULL, sdd.cols = NULL,
  btm = NULL, top = NULL, ylim = NULL, yext = 1.05, min.range = 10,
  max.fold = 5, ratio = 1, gap.width = 1, brk.type = "normal",
  brk.bg = "white", brk.srt = 135, brk.size = 1, brk.col = "black",
  brk.lwd = 1, error.cex = 1, error.lwd = 1, error.col = 1,
  sig.lab = NULL, sig.cex = 1, sig.xpos = 0.5, sig.ypos = -2,
  box.lwd = 1, box.col = 1, box.lty = 1, ...)
}
\arguments{
\item{df}{Long format data.frame, grouped by columns in barplot.}

\item{y.cols}{The index of y columns in df, such y.cols=1:3}

\item{sdu.cols}{If the df data frame contains standard error columns, sdu.cols assigns the column index.}

\item{sdd.cols}{If the upper and lower limits of SD are different, sdd.cols assigns the column index for lower limits.}

\item{btm}{Set the bottom break point value manually. Default is auto.}

\item{top}{Set the top break point value manually. Default is auto.}

\item{ylim}{y limits}

\item{yext}{If ylim is not set, yext sets the y extension.}

\item{min.range}{If the ratio of max y to min y is less than this value, the barplot will not have a gap.}

\item{max.fold}{The max ratio of max(all y) to max(lower part y).}

\item{ratio}{The height ratio of upper to lower part in barplot.}

\item{gap.width}{Relative width of the gap.}

\item{brk.type}{Break type, value should be `normal` or `zigzag`.}

\item{brk.bg}{Background color of the gap.}

\item{brk.srt}{Break line angle if brk.type is `normal`.}

\item{brk.size}{Break line length.}

\item{brk.col}{Break line color.}

\item{brk.lwd}{Break line width.}

\item{error.cex}{cex of error bar.}

\item{error.lwd}{lwd of error bar.}

\item{error.col}{color of error bar.}

\item{sig.lab}{Significance labels vector, the same order as y by columns.}

\item{sig.cex}{cex of sig labels.}

\item{sig.xpos}{Relative x adjustment of sig labels.}

\item{sig.ypos}{Relative y adjustment of sig labels.}

\item{box.lwd}{box line width}

\item{box.col}{box color.}

\item{box.lty}{box line type}

\item{...}{Other parameters passed to barplot function.}
}
\value{
Same as barplot return values.
}
\description{
gap barplot
}
\details{
gap barplot with various options such auto/manual gap interval, errorbar and significance labels.
}
\examples{
## Begin example
library(Xtools)
datax <- na.omit(airquality)[,1:4]
cols <- cm.colors(ncol(datax))
layout(matrix(1:6, ncol=2))
set.seed(0)
for (ndx in 1:6){
    dt <- datax[sample(rownames(datax), 10), ]
    par(mar=c(0.5,2,0.5,0.5))
    brkt <- sample(c('normal', 'zigzag'), 1)
    gap.barplot(dt, col=cols, brk.type=brkt, max.fold=5, ratio=2)
}
## End example
}
\author{
ZG Zhao
}

