% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list.R
\name{get_list}
\alias{get_list}
\title{Get List}
\usage{
get_list(username, bearer_token = Sys.getenv("X_BEARER_TOKEN"))
}
\arguments{
\item{username}{Username of the account that owns the lists}

\item{bearer_token}{A string containing the bearer token for authenticating
with the X API. By default, this argument retrieves the token from the
environment variable \code{X_BEARER_TOKEN} (via \code{Sys.getenv("X_BEARER_TOKEN")}).
Adding your bearer token to your \code{.Renviron} file ensures it is securely
stored and accessible without needing to manually input it for each
session.}
}
\value{
A tibble containing the IDs of the lists and their names, or NULL if none found.
}
\description{
Get a User’s Owned Lists via the \href{https://docs.x.com/x-api/lists/get-a-users-owned-lists}{owned list endpoint}.
}
\examples{
\dontrun{
lists <- get_list(username = "Tesla")
}
}
