% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quote_posts.R
\name{get_quote_posts}
\alias{get_quote_posts}
\title{Retrieve Quote Posts for a Given Post}
\usage{
get_quote_posts(
  post_id,
  max_results = 100,
  exclude = NULL,
  pagination_token = NULL,
  post_fields = c("created_at", "text", "public_metrics", "geo", "attachments",
    "context_annotations", "entities", "lang", "referenced_tweets", "reply_settings",
    "conversation_id", "in_reply_to_user_id", "author_id", "edit_history_tweet_ids",
    "id"),
  user_fields = c("created_at", "description", "protected", "entities", "location",
    "profile_image_url", "public_metrics", "verified", "verified_type"),
  media_fields = c("duration_ms", "height", "width", "preview_image_url", "type", "url",
    "public_metrics", "variants", "media_key"),
  poll_fields = c("end_datetime", "duration_minutes", "options", "voting_status", "id"),
  place_fields = c("contained_within", "country", "country_code", "full_name", "geo",
    "id", "name", "place_type"),
  expansions = c("author_id", "entities.mentions.username",
    "referenced_tweets.id.author_id", "referenced_tweets.id", "in_reply_to_user_id",
    "attachments.media_keys", "attachments.poll_ids"),
  bearer_token = Sys.getenv("X_BEARER_TOKEN")
)
}
\arguments{
\item{post_id}{The ID of the tweet to retrieve quote tweets for.}

\item{max_results}{\code{numeric}; a number between 10 and 100 indicating the
number of posts returned per API call.}

\item{exclude}{A comma-separated list of the types of posts to exclude from
the response (e.g., "retweets", "replies", or "retweets,replies").}

\item{pagination_token}{A string used to navigate backward through result
pages. The X API provides this token when more results are available.
Typically, you won’t need to set this manually as the function handles it,
but you can supply a pagination_token from a previous response to continue
retrieving results beyond the last page, if desired.}

\item{post_fields}{\code{character}, \code{vector}; a list of fields to
return for each post \code{c("created_at", "text", "public_metrics", "geo",
  "attachments","context_annotations", "entities", "lang",
  "referenced_tweets", "reply_settings", "conversation_id",
  "in_reply_to_user_id", "author_id", "edit_history_tweet_ids", "id")}}

\item{user_fields}{\code{character}, \code{vector}; a list of fields to return
for each user \code{c("created_at", "description", "protected", "entities",
"location", "profile_image_url", "public_metrics", "verified",
"verified_type")}}

\item{bearer_token}{A string containing the bearer token for authenticating
with the X API. By default, this argument retrieves the token from the
environment variable \code{X_BEARER_TOKEN} (via \code{Sys.getenv("X_BEARER_TOKEN")}).
Adding your bearer token to your \code{.Renviron} file ensures it is securely
stored and accessible without needing to manually input it for each
session.}
}
\value{
A list containing the API response.
}
\description{
Returns a variety of information about each Post that quotes the Post specified by the requested ID
via the \href{https://docs.x.com/x-api/posts/retrieve-posts-that-quote-a-post}{quote tweets endpoint}.
}
\examples{
\dontrun{
get_quote_posts(tweet_id = "20", max_results = 100)
}
}
