% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_post_count.R
\name{get_recent_post_count}
\alias{get_recent_post_count}
\title{Get Recent Post Count}
\usage{
get_recent_post_count(
  query,
  start_time = NULL,
  end_time = NULL,
  granularity = "hour",
  bearer_token = Sys.getenv("X_BEARER_TOKEN")
)
}
\arguments{
\item{query}{The search to be made on X. You can find ways to build specific queries according to the \href{https://docs.x.com/x-api/posts/search/integrate/build-a-query#types}{X API documentation website}}

\item{start_time}{The earliest date-time from which you want to get posts.}

\item{end_time}{The latest date-time from which you want to get posts.
Provide the value in ISO 8601 format (i.e., \code{YYYY-MM-DDTHH:mm:ssZ}). The
\code{iso_8601()} function will convert a string, date, or date-time object to
the required format (e.g., \code{iso_8601("2024-10-10")}).}

\item{granularity}{The granularity for the search counts results. This takes either the value 'minute', 'hour', or 'day'.}

\item{bearer_token}{A string containing the bearer token for authenticating
with the X API. By default, this argument retrieves the token from the
environment variable \code{X_BEARER_TOKEN} (via \code{Sys.getenv("X_BEARER_TOKEN")}).
Adding your bearer token to your \code{.Renviron} file ensures it is securely
stored and accessible without needing to manually input it for each
session.}
}
\value{
A tibble containing the number of posts.
}
\description{
Returns Post Counts from the last 7 days that match a search query via the \href{https://docs.x.com/x-api/posts/recent-search-counts}{recent posts count endpoint}.
}
\examples{
\dontrun{
tl <- get_recent_post_count("Developers")
}
}
