% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT50s.R
\name{getLT50s}
\alias{getLT50s}
\title{getLT50s() function}
\usage{
getLT50s(df, range.t = NULL, all.info = FALSE)
}
\arguments{
\item{df}{Data frame of electrolyte leakage data (0~1). First column must be "TMP" for temperature.}

\item{range.t}{Temperature range for prediction.}

\item{all.info}{}
}
\value{
If all.info=FALSE (default), return LT50s and their CI95 only. Otherwise return a list consisting original data, LT50s and their CI95, and curve functions.
}
\description{
Calculate LT50s with electrolyte leakage data.
}
\details{
User interface function to calculate LT50s with electrolyte leakage data.
}
\author{
ZG Zhao
}
