% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-LT50s.R
\name{isOverlap}
\alias{isOverlap}
\title{Is two intervals overlaped?}
\usage{
isOverlap(ir1, ir2)
}
\arguments{
\item{ir1}{}

\item{ir2}{}
}
\value{
TRUE/FALSE.
}
\description{
isOverlap
}
\details{
With two given intervals, return TRUE if they are overlap, otherwise return FALSE.
}
\author{
ZG Zhao
}

