% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-LT50s.R
\name{LT50s.shiny.app}
\alias{LT50s.shiny.app}
\title{Using LT50s tools with shiny.}
\usage{
LT50s.shiny.app()
}
\description{
Using LT50s tools with shiny.
}
\examples{
## Begin example
library(Xtools)
datax <- na.omit(airquality)[,1:4]
cols <- cm.colors(ncol(datax))
layout(matrix(1:6, ncol=2))
set.seed(0)
for (ndx in 1:6){
    dt <- datax[sample(rownames(datax), 10), ]
    par(mar=c(0.5,2,0.5,0.5))
    brkt <- sample(c('normal', 'zigzag'), 1)
    gap.barplot(dt, col=cols, brk.type=brkt, max.fold=5, ratio=2)
}
## End example
}
\author{
ZG Zhao
}

