% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-faceting.R
\name{page.faceting}
\alias{page.faceting}
\title{page.facting}
\usage{
page.faceting(n.row, n.col, mar.lft = 0, mar.rht = 0, mar.top = 0, mar.btm = 0)
}
\arguments{
\item{n.row}{}

\item{n.col}{}

\item{mar.lft}{}

\item{mar.rht}{}

\item{mar.top}{}

\item{mar.btm}{}
}
\value{
faceting index
}
\description{
Page faceting
}
\details{
Wraper function for faceting page
}
\author{
ZG Zhao
}
