% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchShow.R
\name{pchShow}
\alias{pchShow}
\title{pchShow}
\usage{
pchShow()
}
\description{
Show pch values and symbols of plot function.
}
\details{
Convenient function to show pch values and symbols of plot function.
}
\author{
ZG Zhao
}
