% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_to_x.R
\name{post_to_x}
\alias{post_to_x}
\title{Post to X}
\usage{
post_to_x(
  text,
  for_super_followers_only = FALSE,
  geo = NULL,
  media = NULL,
  nullcast = FALSE,
  poll = NULL,
  reply = NULL,
  reply_settings = NULL
)
}
\arguments{
\item{text}{The tweet text (max 280 characters)}

\item{for_super_followers_only}{Exclusive Tweet for super followers.}

\item{geo}{Place ID being attached to the Tweet for geo location.}

\item{media}{Media information being attached to created Tweet. This is mutually exclusive from
Quote Tweet Id, Poll, and Card URI.}

\item{nullcast}{Nullcasted (promoted-only) Posts do not appear in the public timeline and are not served to followers.}

\item{poll}{Poll options for a Tweet with a poll. This is mutually exclusive from Media, Quote Tweet Id, and Card URI.}

\item{reply}{Tweet information of the Tweet being replied to.}

\item{reply_settings}{Settings to indicate who can reply to the Tweet.}
}
\description{
Causes the User to create a Post under the authorized account via the \href{https://docs.x.com/x-api/posts/creation-of-a-post}{create a post endpoint}.
}
\examples{
\dontrun{
post_to_x(text = "Hello, world!")
}
}
