% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qPCR.R
\name{qPCR.shiny.app}
\alias{qPCR.shiny.app}
\title{Shiny app for qPCR data processing.}
\usage{
qPCR.shiny.app()
}
\description{
Shiny app for qPCR data processing with BioConductor `HTqPCR` and `ddCt` packages.
}
\examples{
library(Xtools)
qPCR.shiny.app()
}
\author{
ZG Zhao
}

