% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR.R
\name{read.fluoData}
\alias{read.fluoData}
\title{read fluorescence data}
\usage{
read.fluoData(fname, drop = NULL)
}
\arguments{
\item{fname}{Character string: file name.}

\item{drop}{Character vector: Wells to drop, for example, drop=c("A1", "H") will drop A1 and all H wells.}
}
\value{
data.frame.
}
\description{
Read qPCR fluorescence data (Strategene qPCR machine only)
}
\details{
Read qPCR fluo data exported by Strategene MxPro (Export Chart Data::To Text File::Format 1- Vertical grouped).
}
\author{
ZG Zhao
}
