% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repost_to_x.R
\name{repost_to_x}
\alias{repost_to_x}
\title{Repost a Post on X}
\usage{
repost_to_x(tweet_id)
}
\arguments{
\item{tweet_id}{The ID of the post to be reposted.}
}
\description{
Causes the User (in the path) to repost the specified Post. The User in the path must match the User
context authorizing the request. This is done via the \href{https://docs.x.com/x-api/posts/causes-the-user-in-the-path-to-repost-the-specified-post}{retweet endpoint}.
}
\examples{
\dontrun{
repost_to_x(tweet_id = "20")
}
}
