% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_x.R
\name{search_x}
\alias{search_x}
\title{Search X}
\usage{
search_x(
  words_all = "",
  words_exact = "",
  words_any = "",
  words_none = "",
  hashtags = "",
  lang = "en",
  accounts_from = "",
  accounts_to = "",
  accounts_mention = "",
  filter_replies = "",
  filter_links = "",
  min_replies = NULL,
  min_faves = NULL,
  min_reposts = NULL,
  since = NULL,
  until = NULL,
  sort = "top"
)
}
\arguments{
\item{words_all}{A string containing words that must be contained in the query results}

\item{words_exact}{A string containing a phrase that must be contained in the query results}

\item{words_any}{A string containing words or phrases (in quotes) that could be contained in the query results}

\item{words_none}{A string containing words that will not be contained in the query results}

\item{hashtags}{A string containing the hashtags that must be contained in the query result}

\item{lang}{Restrict results by language, which are encoded through their \href{https://localizely.com/iso-639-1-list/}{ISO 639-1 codes}}

\item{accounts_from}{Posts authored by these accounts}

\item{accounts_to}{Posts directed to these accounts}

\item{accounts_mention}{Posts mentioning these accounts}

\item{filter_replies}{Filter replies: "" (show all), "include" (only replies), "exclude" (no replies)}

\item{filter_links}{Filter posts with links: "" (show all), "include" (only links), "exclude" (no links)}

\item{min_replies}{Minimum number of replies the post must have}

\item{min_faves}{Minimum number of likes the post must have}

\item{min_reposts}{Minimum number of reposts (retweets) the post must have}

\item{since}{Start date for post search in YYYY-MM-DD format}

\item{until}{End date for post search in YYYY-MM-DD format}

\item{sort}{Sorting method: "top" (relevancy), "live" (latest)}
}
\value{
A character string containing the full search URL that was opened in the browser.
}
\description{
Search for posts using X's advanced search feature
}
\examples{
\dontrun{
url <- search_x(
words_exact = "full self driving",
words_any = "optimus tesla \"electric vehicle\"",
lang = "en", 
sort = "top", 
since = "2025-01-01",
until = "2025-01-31",
min_faves = 500,
filter_links = "exclude", 
filter_replies = "exclude"
)
}
}
