% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.sigLabs.R
\name{sigLabsDT}
\alias{sigLabsDT}
\title{sigLabsDT}
\usage{
sigLabsDT(dt, col.data = 1:ncol(dt), group.by = NULL)
}
\arguments{
\item{dt}{data.frame, repeats in columns, one row for each treatment/sample}

\item{col.data}{column index of data (repeats)}

\item{group.by}{column index of groups. If set, siglabs will be set independently for each group.}
}
\value{
string vector of significant letter
}
\description{
Set significant labels of multi comparisons with given data.
}
\details{
1. get t-test p-value table of pairwise compairison; 2. set siglabs with function sigLabsPT.
}
\examples{
library("Xtools")
dt <- read.csv(file.path(path.package("Xtools"), "data", "data.measure.csv"))
str(dt)
dx <- cbind(dt[, 1:3], siglabs=sigLabsDT(dt, col.data=4:7, group.by="Plant"))
dx$MM <- apply(dt[,4:7], 1, FUN=mean, na.rm=TRUE)
dx$SD <- apply(dt[,4:7], 1, FUN=sd, na.rm=TRUE)
dx$offset <- max(dx$MM+dx$SD)*0.05
library("ggplot2")
px <- ggplot(dx, aes(x=Leaf, y=MM, fill=Treat)) + theme_bw(base_size = 16)
px + geom_bar(stat="identity", position=position_dodge(.9), width=0.8, color="gray20") +
  geom_errorbar(aes(ymin=MM - SD, ymax=MM + SD), position = position_dodge(.9), width=0.5) +
  geom_text(aes(y=MM + SD + offset, label=siglabs), position=position_dodge(.9), size=5) +
  scale_y_continuous(expand = c(0, 0)) + geom_blank(aes(y = (MM + SD) * 1.15)) +
  facet_wrap(~Plant, nrow=1)
}
\author{
ZG Zhao
}
