% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.sigLabs.R
\name{sigLabsPT}
\alias{sigLabsPT}
\title{sigLabsPT}
\usage{
sigLabsPT(sNames, pTable, scol1 = 1, scol2 = 2, pcol = 3)
}
\arguments{
\item{sNames}{sample names}

\item{pTable}{p-value of statistic test for each pair of samples.}

\item{scol1}{column (name or index) of sample names 1}

\item{scol2}{column (name or index) of sample names 2}

\item{pcol}{column (name or index) of of p-value.}
}
\value{
string vector of significant letter
}
\description{
sigLabsPT: set significant labels for multiple comparisons with given p-value table
}
\details{
Get/set significant labels for multiple comparisons. Samples with any identical letter are not significant different.
}
\examples{
library("Xtools")
pt <- read.csv(file.path(path.package("Xtools"), "data", "data.ptable.csv"))
str(pt)
dt <- subset(pt, subset=Gene=="NR"& Plant=="Arabidopsis")
smps <- sort(unique(c(dt$S1, dt$S2)))
cbind(sample=smps, siglabs=sigLabsPT(smps, dt, scol1=3, scol2=4, pcol=5))
}
\author{
ZG Zhao
}
