% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfollow_user.R
\name{unfollow_user}
\alias{unfollow_user}
\title{Unfollow User}
\usage{
unfollow_user(source_username, target_username)
}
\arguments{
\item{source_username}{Username of account that will unfollow someone.}

\item{target_username}{Username of account that will be unfollowed.}
}
\description{
Causes the source User to unfollow the target User via the \href{https://docs.x.com/x-api/users/unfollow-user}{unfollow user endpoint}.
The source User must match the User context authorizing the request
}
\examples{
\dontrun{
unfollow_user(source_username = "Tesla", target_username = "elonmusk")
}
}
