% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrtags.R
\name{xrtags}
\alias{xrtags}
\title{Modified rtags function.}
\usage{
xrtags(
  path = ".",
  pattern = "\\\\.[RrSs]$",
  recursive = FALSE,
  src = list.files(path = path, pattern = pattern, full.names = TRUE, recursive =
    recursive),
  keep.re = NULL,
  ofile = "",
  append = FALSE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{path}{}

\item{pattern}{}

\item{recursive}{}

\item{src}{}

\item{keep.re}{}

\item{ofile}{}

\item{append}{}

\item{verbose}{}
}
\description{
Modified rtags function.
}
\details{
Modified rtags function.
}
\author{
ZG Zhao
}
