% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gap-barplot.R
\name{gap.barplot}
\alias{gap.barplot}
\title{gap.barplot function}
\usage{
gap.barplot(df, y.cols = 1:ncol(df), sdu.cols = NULL, sdd.cols = NULL,
  btm = NULL, top = NULL, ylim = NULL, yext = 1.05, min.range = 10,
  max.fold = 5, ratio = 1, gap.width = 1, brk.type = "normal",
  brk.bg = "white", brk.srt = 135, brk.size = 1, brk.col = "black",
  brk.lwd = 1, error.cex = 1, error.lwd = 1, error.col = 1,
  sig.lab = NULL, sig.cex = 1, sig.xpos = 0.5, sig.ypos = -2,
  box.lwd = 1, box.col = 1, box.lty = 1, ...)
}
\arguments{
\item{df}{长格式的data.frame，即数据框中每一列为一组绘图数据。}

\item{y.cols}{用做柱形图y值的数据列（序号或名称），一列为一组。如 y.cols=1:3 表示df的1～3列为y值。}

\item{sdu.cols}{df中如果包含误差值，该参数设置误差上限值所在的列，如 sdu.cols=4:6}

\item{sdd.cols}{如果误差值上限和下限不同，可设置该参数，否则默认同sdu.cols参数值。}

\item{btm}{低位断点。如果btm和top均不设置，程序将自动计算和设置断点位置。}

\item{top}{高位断点。}

\item{ylim}{y轴坐标范围}

\item{yext}{y轴延伸比例}

\item{min.range}{自动计算断点的阈值：最大值与最小值的最小比值}

\item{max.fold}{自动计算断点时最大值与下方数据最大值的最大倍数比}

\item{ratio}{断裂后上部与下部y轴长度的比例。}

\item{gap.width}{y轴断裂位置的相对物理宽度（非坐标轴实际刻度）}

\item{brk.type}{断点类型，可设为normal或zigzag}

\item{brk.bg}{断点处的背景颜色}

\item{brk.srt}{断点标记线旋转角度}

\item{brk.size}{断点标记线的大小（长度）}

\item{brk.col}{断点标记线的颜色}

\item{brk.lwd}{断点标记线的线宽}

\item{error.cex}{误差线相对长度，默认为1}

\item{error.lwd}{误差线线宽}

\item{error.col}{误差线颜色}

\item{sig.lab}{显著性标记符号}

\item{sig.cex}{显著性标记符号的字体放大倍数}

\item{sig.xpos}{显著性标记符号的x坐标位移}

\item{sig.ypos}{显著性标记符号的y坐标位移}

\item{box.lwd}{图形区边框线的宽度}

\item{box.col}{图形区边框线的颜色}

\item{box.lty}{图形区边框线的线型}

\item{...}{其他传递给R基本绘图函数barplot的参数}
}
\value{
返回barplot的原始返回值，即柱形图的x坐标
}
\description{
使用R基本绘图函数绘制y轴不连续的柱形图
}
\details{
绘制y轴不连续的柱形图，具有误差线添加功能。断点位置通过btm和top参数设置，如果不设置，函数可自动计算合适的断点位置。
}
\examples{
## Begin example
library(Xtools)
datax <- na.omit(airquality)[,1:4]
cols <- cm.colors(ncol(datax))
layout(matrix(1:6, ncol=2))
set.seed(0)
for (ndx in 1:6){
    dt <- datax[sample(rownames(datax), 10), ]
    par(mar=c(0.5,2,0.5,0.5))
    brkt <- sample(c('normal', 'zigzag'), 1)
    gap.barplot(dt, col=cols, brk.type=brkt, max.fold=5, ratio=2)
}
## End example
}
\author{
ZG Zhao
}

