% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-LT50s.R
\name{LT50s.shiny.app}
\alias{LT50s.shiny.app}
\title{Using LT50s tools with shiny.}
\usage{
LT50s.shiny.app()
}
\description{
Using LT50s tools with shiny.
}
\examples{
## Begin example
library(Xtools)
LT50s.shiny.app()
## End example
}
\author{
ZG Zhao
}

