% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT50s.R
\name{plot.LT50}
\alias{plot.LT50}
\title{Plot LT50 regression curves of given data.}
\usage{
\method{plot}{LT50}(
  dt,
  point.plot = TRUE,
  se.plot = TRUE,
  se.alpha = 0.2,
  labels = NULL,
  show.cols = NA,
  cex.point = 1,
  cex.legend = 1,
  legend.pos = "bottomleft",
  xlab = expression(paste("Temperature (", degree, "C)")),
  ylab = "Electrolyte Leakage (\%)",
  xlim = NULL,
  lwd = 1,
  lty = NULL,
  col = NULL,
  pch = NULL,
  ...
)
}
\arguments{
\item{dt}{The result of getLT50s() with 'all.info=TRUE'.}

\item{point.plot}{If TRUE (default), plot data points.}

\item{se.plot}{If TRUE (default), plot 95\% confidential intervals.}

\item{se.alpha}{Opacity of se shadows.}

\item{labels}{Replace sample names if set.}

\item{show.cols}{Sample columns to show in figure.}

\item{cex.point}{cex for data points.}

\item{cex.legend}{cex for legends.}

\item{legend.pos}{lengend position}

\item{xlab}{Same as 'plot' function.}

\item{ylab}{Same as 'plot' function.}

\item{xlim}{Same as 'plot' function.}

\item{lwd}{Same as 'plot' function.}

\item{lty}{Same as 'plot' function.}

\item{col}{Same as 'plot' function.}

\item{pch}{Same as 'plot' function.}

\item{...}{Other parameters pass to 'plot' function.}
}
\description{
Plot LT50s and return results
}
\details{
Calculation of LT50 with EL data using SSlogis fitting function, returning results of LT50, 95\% CI and R values
}
\author{
ZG Zhao
}
